requirements.txt
streamlit 
pandas
csv
PILLOW
